+++ src/interpreter/lib/std.ts
@@ -517,13 +517,11 @@ export const std: Record<string, Value> = {
 		return OBJ(new Map(obj.value));
 	}),
 
-	/* TODO
 	'Obj:merge': FN_NATIVE(([a, b]) => {
 		assertObject(a);
 		assertObject(b);
-		return OBJ();
+		return OBJ(new Map([...a.value, ...b.value]));
 	}),
-	*/
 	//#endregion
 	
 	//#region Error