+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 バージョン0.16.0に記録漏れがありました。
 >- 関数`Str:from_codepoint` `Str#codepoint_at`を追加
 
+- `package.json`を修正
+
 # 0.17.0
 - 演算子の前後で改行できるように
 ## Breaking changes