+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 # 0.17.0
 - 演算子の前後で改行できるように
+- `package.json`を修正
 ## Breaking changes
 - `match`の構文が変更されました。各行の先頭に`case`または`default`が必要になります。