+++ CHANGELOG.md
@@ -41,6 +41,8 @@
 # 0.13.3
 - 乱数を生成するとき引数の最大値を戻り値に含むように
 
+- インタプリタがreturn文を文字列化できるように
+
 # 0.13.2
 - `Date:year`,`Date:month`,`Date:day`,`Date:hour`,`Date:minute`,`Date:second`に時間数値の引数を渡して時刻指定可能に
 - array.sortとString用比較関数Str:lt, Str:gtの追加