+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 # 0.17.0
 - `package.json`を修正
+- `Error:new`関数でエラー型の値を生成できるように
 - `index_of`の配列版を追加
 - `str.index_of` `arr.index_of`共に第２引数fromIndexを受け付けるように
 - `arr.incl`の引数の型制限を廃止